requirements.txt
Django==5.2.6
psycopg2-binary
django-crispy-forms
crispy-bootstrap4
django-allauth
gunicorn
weasyprint